Moved i18n configuration from the application to its own package and class. MessageSource now autoconfigured via application.properties settings.
