+++ src/main/java/nl/conspect/drivedok/DrivedokApplication.java
@@ -2,55 +2,12 @@ package nl.conspect.drivedok;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.MessageSource;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.support.ResourceBundleMessageSource;
-import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
-import org.springframework.web.servlet.LocaleResolver;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.springframework.web.servlet.i18n.LocaleChangeInterceptor;
-import org.springframework.web.servlet.i18n.SessionLocaleResolver;
-
-import java.util.Locale;
 
 @SpringBootApplication
-public class DrivedokApplication implements WebMvcConfigurer {
+public class DrivedokApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(DrivedokApplication.class, args);
     }
 
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(localeChangeInterceptor());
-    }
-
-    @Bean
-    public LocaleChangeInterceptor localeChangeInterceptor() {
-        return new LocaleChangeInterceptor();
-    }
-
-    @Bean
-    public LocaleResolver localeResolver() {
-        var resolver = new SessionLocaleResolver();
-        resolver.setDefaultLocale(Locale.ENGLISH);
-        return resolver;
-    }
-
-    @Bean
-    public MessageSource messageSource() {
-        var source = new ResourceBundleMessageSource();
-        source.setDefaultEncoding("UTF-8");
-        source.setBasename("i18n/messages");
-        return source;
-    }
-
-    @Bean
-    @Override
-    public LocalValidatorFactoryBean getValidator() {
-        var bean = new LocalValidatorFactoryBean();
-        bean.setValidationMessageSource(messageSource());
-        return bean;
-    }
 }
