+++ src/main/java/nl/conspect/drivedok/utilities/ZoneMapper.java
@@ -6,7 +6,7 @@ import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
 
 @Mapper(componentModel = "spring",
-        unmappedTargetPolicy = ReportingPolicy.IGNORE)
+        unmappedTargetPolicy = ReportingPolicy.WARN)
 public interface ZoneMapper {
 
     Zone dtoToZone(ZoneDto dto);
