+++ src/main/resources/application.properties
@@ -4,3 +4,4 @@ server.error.include-exception=true
 server.error.include-stacktrace=always
 spring.mvc.hiddenmethod.filter.enabled=true
 spring.messages.use-code-as-default-message=true
+spring.messages.basename=i18n/messages
